구매입력 요청근을 클래스로 분리
입력방식이 많아서 인터페이스 사용하여 분리.
