+++ src/main/java/kr/codesquad/view/InputManagerInterface.java
@@ -1,0 +1,7 @@
+package kr.codesquad.view;
+
+import java.util.Optional;
+
+public interface InputManagerInterface<T> {
+	Optional<T> askClient();
+}
